fix: ACNA-1891 refactor console config import to shared module
